[REFACTOR] - Banner, Post ViewModel 분리
