+++ feature/home/src/main/java/com/swmarastro/mykkumi/feature/home/banner/HomeBannerAllFragment.kt
@@ -8,7 +8,6 @@ import androidx.navigation.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.swmarastro.mykkumi.common_ui.base.BaseFragment
 import com.swmarastro.mykkumi.domain.entity.BannerListVO
-import com.swmarastro.mykkumi.feature.home.HomeViewModel
 import com.swmarastro.mykkumi.feature.home.R
 import com.swmarastro.mykkumi.feature.home.databinding.FragmentHomeBannerAllBinding
 import dagger.hilt.android.AndroidEntryPoint
@@ -16,7 +15,7 @@ import dagger.hilt.android.AndroidEntryPoint
 @AndroidEntryPoint
 class HomeBannerAllFragment : BaseFragment<FragmentHomeBannerAllBinding>(R.layout.fragment_home_banner_all) {
 
-    private val viewModel by viewModels<HomeViewModel>({ requireActivity() })
+    private val bannerViewModel by viewModels<HomeBannerViewModel>({ requireActivity() })
     private lateinit var bannerAllAdapter: HomeBannerAllAdapter
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -31,7 +30,7 @@ class HomeBannerAllFragment : BaseFragment<FragmentHomeBannerAllBinding>(R.layou
 
     override suspend fun initView() {
         bind {
-            vm = viewModel
+            bannerVm = bannerViewModel
         }
 
         setBannerAll() // 배너 리스트
@@ -55,15 +54,15 @@ class HomeBannerAllFragment : BaseFragment<FragmentHomeBannerAllBinding>(R.layou
 
     // 배너 내용 세팅
     private suspend fun setBannerAll() {
-        viewModel.setHomeBanner()
-        viewModel.bannerListUiState.collect { response ->
+        bannerViewModel.setHomeBanner()
+        bannerViewModel.bannerListUiState.collect { response ->
             initBannerRecyclerView(response)
         }
     }
 
     // 배너 클릭 -> 배너 상세 페이지 이동
     private fun onClickBannerItem(bannerId: Int) {
-        viewModel.selectHomeBanner(bannerId)
+        bannerViewModel.selectHomeBanner(bannerId)
         view?.findNavController()?.navigate(R.id.action_navigate_fragment_to_home_banner_detail)
     }
